Try harder to detect incorrect labels
The briefness of the instrumentation code made it likely that even the
wrong label would succeed, because the assembler silently assigns
addresses to unrecognised symbols, and sometimes the jumps to those
addresses could accidentally lead to success_code being executed.
Inserting a buffer of infinite-loop instructions between the jump and
the label makes it much less likely that this can ever happen.