Stop inlining all the things
A load of generated assembly will be the same for many scenarios due
to being relative to the stack pointer, so we can just codegen it one
time and call it (without a full stackframe setup cycle, which is
actually one of the most verbose parts of our old codegen).

This is a requirement if we are looking to fit all our instructions into
the 32k provided by the hack hardware.